# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SenaVida API",
        "description": "API REST del backend de SenaVida, plataforma de comunicacion inclusiva para personas sordas en contextos de urgencia medica.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/organizations": {
            "get": {
                "tags": [
                    "Catalogos"
                ],
                "summary": "Listar organizaciones",
                "description": "Devuelve todas las organizaciones de salud registradas en el sistema.",
                "operationId": "212e7ae65bac93ef8fe7a173dd8c21ed",
                "responses": {
                    "200": {
                        "description": "Listado de organizaciones"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catalogos"
                ],
                "summary": "Crear organizacion",
                "description": "Crea una nueva organizacion de salud. Solo el super_admin puede hacerlo.",
                "operationId": "242e48fe4b9103f0f1d4b5a5859f95fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Servicio de Salud Metropolitano"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organizacion creada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Solo super_admin puede crear organizaciones"
                    },
                    "422": {
                        "description": "Datos invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/units": {
            "get": {
                "tags": [
                    "Catalogos"
                ],
                "summary": "Listar unidades",
                "description": "Devuelve las unidades del sistema. Si se envia el parametro healthCenterId, filtra solo las de ese centro.",
                "operationId": "f1ee53f15b9381a8d3f28d8b6d5fa4cc",
                "parameters": [
                    {
                        "name": "healthCenterId",
                        "in": "query",
                        "description": "UUID del centro de salud para filtrar sus unidades (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de unidades"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catalogos"
                ],
                "summary": "Crear unidad",
                "description": "Crea una unidad dentro de un centro de salud. El super_admin puede crearla en cualquier centro; el admin_institucional solo en el suyo.",
                "operationId": "60839c16039ec69ff8377a4da20528af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "healthCenterId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Urgencia Adulto"
                                    },
                                    "healthCenterId": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Unidad creada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso, o fuera del propio centro"
                    },
                    "422": {
                        "description": "Datos invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Bearer obtenido en /auth/login",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}